# chore: Reimplement ShuffleWriterExec using interleave_record_batch