--- a/native/core/src/execution/shuffle/codec.rs
+++ b/native/core/src/execution/shuffle/codec.rs
@@ -750,16 +750,15 @@ mod test {
     }
 }
 
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone)]
 pub enum CompressionCodec {
-    #[default]
     None,
     Lz4Frame,
     Zstd(i32),
     Snappy,
 }
 
-#[derive(Debug, Clone, Default)]
+#[derive(Clone)]
 pub struct ShuffleBlockWriter {
     fast_encoding: bool,
     codec: CompressionCodec,