<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #1235, #1449 and partially #1446, #1453.

## Rationale for this change

Using `interleave_record_batch` instead of maintaining array builders for each partition improves memory utilization and reduces the possibility of excessive spilling. For instance, the test case `test_large_number_of_partitions_spilling` spills 3332 times before this change, now it spills only 3 times.

Switching from array builders to `interleave_record_batch` also reduces the amount of code we have to maintain. Now we can completely get rid of `builder.rs`, and the logic for handle partitioning and spilling is easier to to track than before.

## What changes are included in this PR?

Reimplement ShuffleWriteExec using `interleave_record_batch` and get rid of the old builder-based approach.

## How are these changes tested?

* Passing existing tests
* Tested on TPC-H SF=100