@andygrove 2025-03-12
I ran a quick TPC-H benchmark. There is a slight regression in performance, with this PR taking 323 seconds compared to 295 seconds in the main branch. However, we have the advantage of less memory overhead and support for complex types (in theory).