@andygrove 2025-03-12
> The failure in java-native_iceberg_compat test is caused by a bug in the [fast shuffle writer](https://github.com/apache/datafusion-comet/blob/58a36b62004c1e300d7f3113d9a64713a94c0db9/native/core/src/execution/shuffle/codec.rs#L302-L304). The null buffers written for sliced arrays are not correct. I'll submit a dedicated patch to fix it later.

There is now an option in the Arrow IPC reader to disable re-validating the data and we may be able to remove our custom encoding if that solves the performance issues we saw previously.