@mbutrovich 2025-03-21
Dumb question: is `take_record_batch` an option here (it looks like interleave can actually coalesce a vector or `RecordBatch`. What would the performance implication be of `take` vs. `interleave`?