@Kontinuation 2025-03-21
> Dumb question: is `take_record_batch` an option here (it looks like interleave can actually coalesce a vector of `RecordBatch`. What would the performance implication be of `take` vs. `interleave`?

The `take_record_batch` function is likely to result in numerous small memory allocations for each input batch, which may lead to poorer performance compared to `interleave_record_batch`. However, it's essential to prototype both implementations to fully comprehend their performance characteristics.